角色列表

+++ service/admin/admin-service/src/main/java/com/jkm/hss/admin/dao/AdminUserDao.java
@@ -79,7 +79,7 @@ public interface AdminUserDao {
      * @param adminUserListRequest
      * @return
      */
-    int selectAdminUserCountByPageParams(AdminUserListRequest adminUserListRequest);
+    long selectAdminUserCountByPageParams(AdminUserListRequest adminUserListRequest);
 
     /**
      * 分页查询员工列表
